ib-visual-backend/requirements.txt
ib_insync
websockets
influxdb
requests
TA-Lib
python-dateutil